refactor: relatedArticleTabs 리네이밍(ArticleTabs) 및 폴더 위치 shared로 변경(#141)
