+++ src/shared/components/articleTabs/articleTabs.css.ts
